Use Immutable collection adapters from kotlinx.collections.immutables
